+++ tests/test_visualization_aids.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
 import matplotlib
+import pytest
 
 matplotlib.use('agg')
 import numpy
@@ -9,12 +10,17 @@ from matplotlib import pyplot
 from sklearn import datasets
 from sklearn.tree import DecisionTreeClassifier
 
-from ds_utils.visualization_aids import draw_tree, visualize_features, draw_dot_data, visualize_correlations
+from ds_utils.visualization_aids import draw_tree, visualize_features, draw_dot_data, visualize_correlations, \
+    plot_features_relationship
 from tests.utils import compare_images_paths
 
 iris = datasets.load_iris()
 x = iris.data
 y = iris.target
+
+data_1M = pandas.read_csv(Path(__file__).parents[0].joinpath("resources").joinpath("data.1M.zip"), compression='zip')
+loan_data = pandas.read_csv(Path(__file__).parents[0].joinpath("resources").joinpath("loan_final313.csv"),
+                            encoding="latin1", parse_dates=["issue_d"]).drop("id", axis=1)
 
 Path(__file__).parents[0].absolute().joinpath("result_images").mkdir(exist_ok=True)
 Path(__file__).parents[0].absolute().joinpath("result_images").joinpath("test_visualization_aids").mkdir(
@@ -113,11 +119,7 @@ def test_draw_dot_data_exist_ax():
 
 
 def test_visualize_features():
-    file_path = Path(__file__).parents[0].joinpath("resources").joinpath("loan_final313.csv")
-    loan_frame = pandas.read_csv(file_path, encoding="latin1", parse_dates=["issue_d"])
-    loan_frame = loan_frame.drop("id", axis=1)
-
-    visualize_features(loan_frame)
+    visualize_features(loan_data)
 
     result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
         "test_visualization_aids").joinpath("test_visualize_features.png")
@@ -143,15 +145,13 @@ def test_visualize_features_list_of_features():
 
 
 def test_visualize_features_remove_na():
-    file_path = Path(__file__).parents[0].joinpath("resources").joinpath("loan_final313.csv")
-    loan_frame = pandas.read_csv(file_path, encoding="latin1", parse_dates=["issue_d"])
-    loan_frame = loan_frame.drop("id", axis=1)
-    loan_frame = loan_frame.sample(1000, random_state=0)
-    loan_frame = pandas.concat(
-        [loan_frame, pandas.DataFrame([[numpy.nan] * len(loan_frame.columns)] * 250, columns=loan_frame.columns)],
+    loan_data_dup = loan_data.sample(1000, random_state=0)
+    loan_data_dup = pandas.concat(
+        [loan_data_dup,
+         pandas.DataFrame([[numpy.nan] * len(loan_data_dup.columns)] * 250, columns=loan_data_dup.columns)],
         ignore_index=True).sample(frac=1, random_state=0)
 
-    visualize_features(loan_frame, remove_na=True)
+    visualize_features(loan_data_dup, remove_na=True)
 
     result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
         "test_visualization_aids").joinpath("test_visualize_features_remove_na.png")
@@ -164,24 +164,7 @@ def test_visualize_features_remove_na():
 
 
 def test_visualize_correlations():
-    file_path = Path(__file__).parents[0].joinpath("resources").joinpath("data.1M.zip")
-    data = pandas.read_csv(file_path, compression='zip', index_col=0)
-    data_minimal = data[['days_since_first', 'positives', 'scan_date', 'times_submitted', 'unique_sources',
-                         'additional_info.exiftool.CodeSize', 'additional_info.exiftool.FileType',
-                         'additional_info.exiftool.InitializedDataSize',
-                         'additional_info.exiftool.UninitializedDataSize',
-                         'additional_info.exiftool.PEType', 'size', 'network_activity']].copy()
-    data_minimal["network_activity"] = data_minimal["network_activity"].astype("bool")
-    data_minimal = data_minimal.rename(
-        {"days_since_first": "x1", "positives": "x2", "scan_date": "x3", "times_submitted": "x4",
-         "unique_sources": "x5", "additional_info.exiftool.CodeSize": "x6",
-         "additional_info.exiftool.FileType": "x7",
-         "additional_info.exiftool.InitializedDataSize": "x8",
-         "additional_info.exiftool.UninitializedDataSize": "x9",
-         "additional_info.exiftool.PEType": "x10", "size": "x11",
-         "network_activity": "x12"}, axis="columns")
-
-    visualize_correlations(data_minimal)
+    visualize_correlations(data_1M)
 
     result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
         "test_visualization_aids").joinpath("test_visualize_correlations.png")
@@ -194,29 +177,12 @@ def test_visualize_correlations():
 
 
 def test_visualize_correlations_exist_ax():
-    file_path = Path(__file__).parents[0].joinpath("resources").joinpath("data.1M.zip")
-    data = pandas.read_csv(file_path, compression='zip', index_col=0)
-    data_minimal = data[['days_since_first', 'positives', 'scan_date', 'times_submitted', 'unique_sources',
-                         'additional_info.exiftool.CodeSize', 'additional_info.exiftool.FileType',
-                         'additional_info.exiftool.InitializedDataSize',
-                         'additional_info.exiftool.UninitializedDataSize',
-                         'additional_info.exiftool.PEType', 'size', 'network_activity']].copy()
-    data_minimal["network_activity"] = data_minimal["network_activity"].astype("bool")
-    data_minimal = data_minimal.rename(
-        {"days_since_first": "x1", "positives": "x2", "scan_date": "x3", "times_submitted": "x4",
-         "unique_sources": "x5", "additional_info.exiftool.CodeSize": "x6",
-         "additional_info.exiftool.FileType": "x7",
-         "additional_info.exiftool.InitializedDataSize": "x8",
-         "additional_info.exiftool.UninitializedDataSize": "x9",
-         "additional_info.exiftool.PEType": "x10", "size": "x11",
-         "network_activity": "x12"}, axis="columns")
-
     pyplot.figure()
     ax = pyplot.gca()
 
     ax.set_title("My ax")
 
-    visualize_correlations(data_minimal, ax=ax)
+    visualize_correlations(data_1M, ax=ax)
 
     result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
         "test_visualization_aids").joinpath("test_visualize_correlations_exist_ax.png")
@@ -226,3 +192,106 @@ def test_visualize_correlations_exist_ax():
     baseline_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
         "test_visualization_aids").joinpath("test_visualize_correlations_exist_ax.png")
     compare_images_paths(str(baseline_path), str(result_path))
+
+
+def test_plot_relationship_between_features_both_numeric():
+    plot_features_relationship("x4", "x5", data_1M)
+
+    result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_both_numeric.png")
+    pyplot.savefig(str(result_path))
+
+    baseline_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_both_numeric.png")
+    compare_images_paths(str(baseline_path), str(result_path))
+
+
+def test_plot_relationship_between_features_both_numeric_exist_ax():
+    pyplot.figure()
+    ax = pyplot.gca()
+
+    ax.set_title("My ax")
+
+    plot_features_relationship("x4", "x5", data_1M, ax=ax)
+
+    result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_both_numeric_exist_ax.png")
+    pyplot.savefig(str(result_path))
+
+    baseline_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_both_numeric_exist_ax.png")
+    compare_images_paths(str(baseline_path), str(result_path))
+
+
+def test_plot_relationship_between_features_numeric_categorical():
+    plot_features_relationship("x1", "x7", data_1M)
+
+    result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_numeric_categorical.png")
+    pyplot.gcf().set_size_inches(14, 9)
+    pyplot.savefig(str(result_path))
+
+    baseline_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_numeric_categorical.png")
+    compare_images_paths(str(baseline_path), str(result_path))
+
+
+def test_plot_relationship_between_features_numeric_categorical_reverse():
+    plot_features_relationship("x7", "x1", data_1M)
+
+    result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_numeric_categorical_reverse.png")
+    pyplot.gcf().set_size_inches(14, 9)
+    pyplot.savefig(str(result_path))
+
+    baseline_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_numeric_categorical_reverse.png")
+    compare_images_paths(str(baseline_path), str(result_path))
+
+
+def test_plot_relationship_between_features_numeric_boolean():
+    plot_features_relationship("x1", "x12", data_1M)
+
+    result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_numeric_boolean.png")
+    pyplot.gcf().set_size_inches(8, 7)
+    pyplot.savefig(str(result_path))
+
+    baseline_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_numeric_boolean.png")
+    compare_images_paths(str(baseline_path), str(result_path))
+
+
+def test_plot_relationship_between_features_both_categorical():
+    plot_features_relationship("x7", "x10", data_1M)
+
+    result_path = Path(__file__).parents[0].absolute().joinpath("result_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_both_categorical.png")
+    pyplot.gcf().set_size_inches(9, 5)
+    pyplot.savefig(str(result_path))
+
+    baseline_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
+        "test_visualization_aids").joinpath("test_plot_relationship_between_features_both_categorical.png")
+    compare_images_paths(str(baseline_path), str(result_path))
+
+
+@pytest.mark.skip()
+def test_loop_plot_features_relationship_example():
+    figure, axes = pyplot.subplots(6, 2)
+    axes = axes.flatten()
+    figure.set_size_inches(16, 25)
+
+    feature_1 = "x1"
+    other_features = ["x2", "x3", "x4", "x5", "x6", "x7", "x8", "x9", "x10", "x11", "x12"]
+
+    for i in range(0, len(other_features)):
+        axes[i].set_title(f"{feature_1} vs. {other_features[i]}")
+        plot_features_relationship(feature_1, other_features[i], data_1M, ax=axes[i])
+
+    figure.delaxes(axes[11])
+    figure.subplots_adjust(hspace=0.7)
+    pyplot.show()
+
+    result_path = Path(__file__).parents[0].absolute().joinpath("baseline_images").joinpath(
+        "test_visualization_aids").joinpath("loop_plot_features_relationship_example.png")
+    pyplot.savefig(str(result_path))